feat: redirect users with no workspaces to onboarding page
