+++ apps/web/src/app/sign-in/[[...sign-in]]/page.tsx
@@ -4,7 +4,7 @@ import { SignIn } from "@clerk/nextjs";
 export default function Page() {
   return (
     <div className="flex min-h-screen items-center justify-center p-6">
-      <SignIn afterSignInUrl="/workspaces"/>
+      <SignIn afterSignInUrl="/workspaces/onboarding"/>
     </div>
   );
 }
