+++ apps/web/src/app/sign-up/[[...sign-up]]/page.tsx
@@ -4,7 +4,7 @@ import { SignUp } from "@clerk/nextjs";
 export default function Page() {
   return (
     <div className="flex min-h-screen items-center justify-center p-6">
-      <SignUp afterSignInUrl="/workspaces"/>
+      <SignUp afterSignInUrl="/workspaces/onboarding"/>
     </div>
   );
 }
